test(mfsdp): separate memory tests
