+++ tests/unit_tests/distributed/mfsdp_v2/test_fully_shard.py
@@ -86,18 +86,6 @@ class MultiChildModel(nn.Module):
         return x
 
 
-class ElementwiseModel(nn.Module):
-    """Small activation path over a large FSDP-managed weight."""
-
-    def __init__(self, dim: int) -> None:
-        super().__init__()
-        self.weight = nn.Parameter(torch.ones(dim, dim))
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        """Apply the first weight row to an activation tensor."""
-        return torch.relu(x + self.weight[0])
-
-
 class TiedLM(nn.Module):
     """Tiny language model with shared input and output embedding weights."""
 
@@ -170,10 +158,6 @@ def _hfsdp_placements() -> Placements:
         gradient=[Partial(dist.ReduceOp.AVG), Flat()],
         optimizer=[Flat(), Flat()],
     )
-
-
-def _mb(num_bytes: int) -> str:
-    return f"{num_bytes / 1024**2:.2f} MB"
 
 
 # CPU ops that a device event chains up to via cpu_parent, used to attribute the device
@@ -603,172 +587,6 @@ def test_tied_child_parameters_allocate_one_physical_weight(distributed_setup):
     assert len(list(model.parameters())) == 1
 
 
-@pytest.mark.parametrize("main_params_dtype", [torch.bfloat16, torch.float32])
-def test_persistent_sharded_storage(distributed_setup, main_params_dtype):
-    """FSDP should retain only its sharded weights and gradients at rest."""
-    rank = distributed_setup.rank
-    world_size = distributed_setup.world_size
-    device = distributed_setup.device
-    if world_size < 2:
-        pytest.skip("This test requires at least 2 ranks.")
-
-    mesh = init_device_mesh(device.type, (world_size,))
-    dim = 4096
-    dtype = torch.bfloat16
-    model = MultiChildModel(dim=dim, num_children=8).to(dtype=dtype)
-    placements = _flat_placements()
-    policy = MixedPrecisionPolicy(main_params_dtype=main_params_dtype)
-    allocated_before = torch.cuda.memory_allocated(device)
-    with fully_shard_context(device=device):
-        for layer in model.layers:
-            fully_shard(layer, mesh=mesh, placements=placements, mixed_precision_policy=policy)
-        fully_shard(model, mesh=mesh, placements=placements, mixed_precision_policy=policy)
-
-    child_weight_nbytes = dim * dim * torch.empty((), dtype=dtype).element_size()
-    persistent_allocated = torch.cuda.memory_allocated(device) - allocated_before
-    if main_params_dtype == dtype:
-        # Model and main weights alias, leaving only one BF16 weight buffer and one
-        # BF16 main-gradient buffer per child.
-        assert all(
-            group.model_weight is group.main_weight
-            for layer in model.layers
-            for group in layer.parameter_groups
-        )
-        expected_per_child_nbytes = 2 * child_weight_nbytes
-    else:
-        # FP32 main weights require a distinct buffer in addition to the BF16 model
-        # weight and BF16 main-gradient buffers.
-        main_weight_nbytes = dim * dim * torch.empty((), dtype=main_params_dtype).element_size()
-        expected_per_child_nbytes = 2 * child_weight_nbytes + main_weight_nbytes
-
-    # All persistent buffers are sharded over the data-parallel group. Small bookkeeping
-    # allocations stay below 1 MiB.
-    expected_persistent_nbytes = len(model.layers) * expected_per_child_nbytes // world_size
-    assert (
-        expected_persistent_nbytes <= persistent_allocated < expected_persistent_nbytes + 1024**2
-    ), (
-        "FSDP persistent memory does not match its sharded weight and gradient storage: "
-        f"rank={rank}, persistent_allocated={_mb(persistent_allocated)}, "
-        f"expected={_mb(expected_persistent_nbytes)}"
-    )
-
-
-def test_training_step_peak_memory_bounds_full_size_buffers(distributed_setup):
-    """A training step should stay below five full-size child buffers."""
-    rank = distributed_setup.rank
-    world_size = distributed_setup.world_size
-    device = distributed_setup.device
-    if world_size < 2:
-        pytest.skip("This test requires at least 2 ranks.")
-
-    mesh = init_device_mesh(device.type, (world_size,))
-    dim = 4096
-    dtype = torch.bfloat16
-    model = MultiChildModel(dim=dim, num_children=8).to(dtype=dtype)
-    placements = _flat_placements()
-    policy = MixedPrecisionPolicy(main_params_dtype=dtype, main_grads_dtype=dtype)
-    with fully_shard_context(device=device):
-        for layer in model.layers:
-            fully_shard(layer, mesh=mesh, placements=placements, mixed_precision_policy=policy)
-        fully_shard(model, mesh=mesh, placements=placements, mixed_precision_policy=policy)
-
-    optimizer = torch.optim.SGD(model.parameters(), lr=0.01)
-    fully_shard_optimizer(optimizer)
-    x = torch.randn(2, dim, device=device, dtype=dtype)
-
-    def train_step() -> None:
-        optimizer.zero_grad(set_to_none=True)
-        model(x).float().sum().backward()
-        optimizer.step()
-
-    child_weight_nbytes = dim * dim * torch.empty((), dtype=dtype).element_size()
-    resting_allocated = torch.cuda.memory_allocated(device)
-    torch.cuda.reset_peak_memory_stats(device)
-    train_step()
-    peak_delta = torch.cuda.max_memory_allocated(device) - resting_allocated
-
-    # Backward keeps the current child and one prefetched child unsharded. The current
-    # child also has a full wgrad until it is copied into a full reduce-scatter input,
-    # for a four-full-child-buffer peak. Allow one additional buffer for cuBLAS
-    # workspace, allocator granularity, and small temporaries.
-    bound_nbytes = (4 + 1) * child_weight_nbytes
-
-    assert peak_delta < bound_nbytes, (
-        "FSDP training-step peak memory exceeded the full-size-buffer bound: "
-        f"rank={rank}, peak_delta={_mb(peak_delta)}, "
-        f"five_full_child_buffers={_mb(bound_nbytes)}"
-    )
-
-
-def test_deleted_model_releases_fsdp_storage(distributed_setup):
-    """Deleting an FSDP model should release its persistent storage."""
-    world_size = distributed_setup.world_size
-    device = distributed_setup.device
-
-    mesh = init_device_mesh(device.type, (world_size,))
-    # Earlier tests may retain process-global CUDA allocations such as the
-    # CuBLAS workspace. Capture them before creating this model, so the test
-    # only detects storage retained by the deleted FSDP model itself.
-    allocated_before = torch.cuda.memory_allocated(device)
-    model = ElementwiseModel(dim=8192).to(dtype=torch.bfloat16, device=device)
-    with fully_shard_context(device=device):
-        fully_shard(model, mesh=mesh, placements=_flat_placements())
-
-    x = torch.ones(1, 8192, dtype=torch.bfloat16, device=device)
-    output = model(x)
-    del output, x, model
-
-    assert torch.cuda.memory_allocated(device) - allocated_before < 1024**2
-
-
-def test_fully_shard_returns_to_resting_memory(distributed_setup):
-    """Fully-sharded temporary storage should be released after forward and backward."""
-    rank = distributed_setup.rank
-    world_size = distributed_setup.world_size
-    device = distributed_setup.device
-    if world_size < 2:
-        pytest.skip("This test requires at least 2 ranks.")
-
-    mesh = init_device_mesh(device.type, (world_size,))
-    dim = 4096
-    dtype = torch.bfloat16
-    model = MultiChildModel(dim=dim, num_children=2).to(dtype=dtype, device=device)
-    placements = _flat_placements()
-    policy = MixedPrecisionPolicy(main_params_dtype=dtype, main_grads_dtype=dtype)
-    with fully_shard_context(device=device):
-        for layer in model.layers:
-            fully_shard(layer, mesh=mesh, placements=placements, mixed_precision_policy=policy)
-        fully_shard(model, mesh=mesh, placements=placements, mixed_precision_policy=policy)
-
-    x = torch.randn(2, dim, device=device, dtype=dtype)
-
-    def clear_cublas_workspaces_and_get_allocated_memory() -> int:
-        # PyTorch retains a cuBLAS workspace for each handle/stream pair. Clear those
-        # library caches so this measurement isolates FSDP-managed storage.
-        torch._C._cuda_clearCublasWorkspaces()
-        return torch.cuda.memory_allocated(device)
-
-    resting_allocated = clear_cublas_workspaces_and_get_allocated_memory()
-
-    def assert_returns_to_resting_memory(phase: str) -> None:
-        extra_allocated = clear_cublas_workspaces_and_get_allocated_memory() - resting_allocated
-        # The live output, activations, and root-owned bias gradient are small; unsharded
-        # parameter storage must be released.
-        assert extra_allocated < 1024**2, (
-            f"Fully-sharded storage did not return to resting memory after {phase}: "
-            f"rank={rank}, extra_allocated={_mb(extra_allocated)}, "
-            "max_extra_allocated=1.00 MB"
-        )
-
-    output = model(x)
-    assert_returns_to_resting_memory("forward")
-
-    loss = output.float().square().mean()
-    loss.backward()
-    del loss, output
-    assert_returns_to_resting_memory("backward")
-
-
 def test_parameterless_parent_with_child_modules_trains(distributed_setup):
     """A parent with no unowned parameters should still root trainable child FsdpModules."""
     world_size = distributed_setup.world_size
@@ -1049,63 +867,3 @@ def test_non_leaf_parameter_view_survives_storage_resize(distributed_setup):
     assert group.main_grad is not None
     assert group._unsharded_model_weight is not None
     assert group._unsharded_model_weight.local_buffer.untyped_storage().nbytes() == 0
-
-
-def test_fully_shard_reduces_peak_training_memory(distributed_setup):
-    """Per-layer FSDP should reduce peak CUDA memory during training."""
-    rank = distributed_setup.rank
-    world_size = distributed_setup.world_size
-    device = distributed_setup.device
-    if world_size < 2:
-        pytest.skip("This test requires at least 2 ranks.")
-    mesh = init_device_mesh(device.type, (world_size,))
-    dim = 1024
-    layers = 16
-    batch = 8
-    steps = 2
-    dtype = torch.bfloat16
-
-    def train_steps(model: nn.Module, optimizer: torch.optim.Optimizer, x: torch.Tensor) -> None:
-        for _ in range(steps):
-            optimizer.zero_grad(set_to_none=True)
-            model(x).sum().backward()
-            optimizer.step()
-
-    torch.manual_seed(4321)
-    baseline = nn.Sequential(*[nn.Linear(dim, dim, dtype=dtype) for _ in range(layers)]).to(device)
-    baseline_optimizer = torch.optim.AdamW(baseline.parameters(), lr=0.01)
-    x = torch.randn(batch, dim, device=device, dtype=dtype)
-    torch.cuda.reset_peak_memory_stats(device)
-    train_steps(baseline, baseline_optimizer, x)
-    baseline_peak = torch.cuda.max_memory_allocated(device)
-
-    del baseline_optimizer
-    del baseline
-    del x
-
-    torch.manual_seed(4321)
-    model = nn.Sequential(*[nn.Linear(dim, dim, dtype=dtype) for _ in range(layers)]).to(device)
-    with fully_shard_context(device=device):
-        for layer in model:
-            fully_shard(
-                layer,
-                mesh=mesh,
-                placements=_flat_placements(),
-                mixed_precision_policy=MixedPrecisionPolicy(
-                    main_params_dtype=dtype, main_grads_dtype=dtype
-                ),
-            )
-    optimizer = torch.optim.AdamW(model.parameters(), lr=0.01)
-
-    x = torch.randn(batch, dim, device=device, dtype=dtype)
-    torch.cuda.reset_peak_memory_stats(device)
-    train_steps(model, optimizer, x)
-    sharded_peak = torch.cuda.max_memory_allocated(device)
-    logger.info(
-        "FSDP peak memory: rank=%s, baseline=%s, sharded=%s",
-        rank,
-        _mb(baseline_peak),
-        _mb(sharded_peak),
-    )
-
-    assert sharded_peak < baseline_peak
